+++ src/main/java/com/mdmc/posofmyheart/application/dtos/OrderRequest.java
@@ -9,7 +9,7 @@ import java.math.BigDecimal;
 import java.util.List;
 
 public record OrderRequest(
-        @NotNull @Positive Integer idPaymentMethod,
+        @NotNull @Positive Long idPaymentMethod,
         @NotNull @Positive BigDecimal amount,
         @Size(max = 100) String comment,
         @NotEmpty List<OrderItemRequest> items
