+++ src/main/java/com/mdmc/posofmyheart/infrastructure/persistence/repositories/PaymentMethodRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface PaymentMethodRepository extends JpaRepository<PaymentMethodEntity, Integer> {
+public interface PaymentMethodRepository extends JpaRepository<PaymentMethodEntity, Long> {
     List<PaymentMethodEntity> findByActiveTrue();
 }
