+++ src/main/java/com/mdmc/posofmyheart/infrastructure/persistence/repositories/VariantRepository.java
@@ -3,5 +3,5 @@ package com.mdmc.posofmyheart.infrastructure.persistence.repositories;
 import com.mdmc.posofmyheart.infrastructure.persistence.entities.ProductVariantEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VariantRepository extends JpaRepository<ProductVariantEntity, Integer> {
+public interface VariantRepository extends JpaRepository<ProductVariantEntity, Long> {
 }
